+++ tsMuxer/mpegAudioStreamReader.h
@@ -12,7 +12,7 @@ class MpegAudioStreamReader final : public SimplePacketizerReader, MP3Codec
     int getTSDescriptor(uint8_t* dstBuff, bool blurayMode, bool hdmvDescriptors) override;
     int getLayer() const { return m_layer; }
     int getFreq() override { return m_sample_rate; }
-    int getChannels() override { return 2; }
+    uint8_t getChannels() override { return 2; }
 
    protected:
     int getHeaderLen() override { return MPEG_AUDIO_HEADER_SIZE; }
